backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.1
tiktoken==0.9.0
nltk==3.9.1
openai==1.75.0
rank-bm25==0.2.2
numpy==2.2.4
python-dotenv==1.1.0
sentence-transformers>=3.0.0
cohere>=5.0.0
anthropic>=0.50.0
